+++ src/pydocstyle/utils.py
@@ -2,7 +2,7 @@
 import logging
 
 
-__version__ = '1.1.1'
+__version__ = '2.0.0rc'
 log = logging.getLogger(__name__)