+++ src/tests/test_cases/test.py
@@ -1,5 +1,7 @@
 # encoding: utf-8
 # No docstring, so we can test D100
+from functools import wraps
+import os
 import sys
 from .expected import Expectation
 
@@ -362,5 +364,11 @@ def docstring_ignore_violations_of_pydocstyle_D400_and_PEP8_E501_but_catch_D401(
     """Runs something"""
     pass
 
-expect(__file__ if __file__[-1] != 'c' else __file__[:-1],
+
+@wraps(docstring_bad_ignore_one)
+def bad_decorated_function():
+    """Bad (E501) but decorated"""
+    pass
+
+expect(os.path.normcase(__file__ if __file__[-1] != 'c' else __file__[:-1]),
        'D100: Missing docstring in public module')