+++ docs/release_notes.rst
@@ -4,13 +4,30 @@ Release Notes
 **pydocstyle** version numbers follow the
 `Semantic Versioning <http://semver.org/>`_ specification.
 
+Current Development Version
+---------------------------
+
+Major Updates
+
+* Support for Python 2.6 has been dropped (#206, #217).
+* Support for the ``pep257`` console script has been dropped. Only the
+  ``pydocstyle`` console script should be used (#216, #218).
+
+New Features
+
+* Decorator-based skipping via ``--ignore-decorators`` has been added (#204).
+
+Bug Fixes
+
+* Made parser more robust to bad source files (#168, #214)
+
 1.1.1 - October 4th, 2016
 -------------------------
 
 Bug Fixes
 
-* Fixed an issue where the `flake8-docstrings` failed when accessing some
-  public API from `pydocstyle`.
+* Fixed an issue where the ``flake8-docstrings`` failed when accessing some
+  public API from ``pydocstyle``.
 
 
 1.1.0 - September 29th, 2016